tests: create a DCAF context to be used in unit tests
The new function dcaf_context() implements a singleton that creates
a new dcaf_context_t with a default configuration on first use.

A test was added in test_ticket.cc that uses this context to set
the CoAP context for a made-up session object that is passed to
dcaf_set_ticket_grant().