dcaf_optlist.[hc]: Set DCAF context options
The new function dcaf_set_option() provides an interface to
control context-specific options during runtime. The only option
defined so far is DCAF_OPTION_TIMEOUT. This option denotes the
timeout in milliseconds until a transaction is considered failed
(default: 90000).

Usage:

   dcaf_option_t timeout = { .v.uint = 120000 };
   dcaf_set_option(context, DCAF_OPTION_TIMEOUT, &timeout);

This example sets the timeout for context to 120000 ms (2 minutes).