dcaf-client.c: split verbosity specification
The argument -v now takes up to three numeric values separated by
comma. The first number is the DCAF verbosity, the second number the
verbosity of the CoAP library, and the third number the verbosity of
the DTLS library. Numbers from the end may be left out to use the
default log level.