dcaf.c: TLS support
When TCP/TLS is supported in the underlying CoAP library, TCP and
TLS endpoints are created at startup. The transport used for the
initial communication attempt is stored by dcaf_send_request_uri().
After retrieval of an access token, the secure variant of this
transport protocol is used for the subsequent session establishment
attempt.