[AM] config_parser.cc: Make transport address configurable
The YAML structure for specifying endpoints now contains a list
of interfaces and optional transport addresses. For example, to
specify an interface aaaa::1 with the transport addresses 7744
for UDP and TCP, and 7745 for DTLS and TLS, the following would
be included in the configuration file:

endpoints:
  - interface:
      address: "aaaa::1"
      udp: 7744
      dtls: 7745
      tcp: 7744
      tls: 7745

When one of {udp,tcp,dtls,tls} is provided, no default port numbers
will be used. Endpoints for protocols that are not specified will not
be created. For example, to offer only TCP-based endpoints, the
following can be specified:

endpoints:
  - interface:
      address: "aaaa::1"
      tcp: 12000
      tls: 12001

The following selects the default transport endpoints for all supported
protocols for the localhost interface:

endpoints:
  - interface:
      address: "::1"